Track turns with a MovementComplete event.

+++ src/grid.rs
@@ -11,6 +11,7 @@ pub struct GridPlugin;
 impl Plugin for GridPlugin {
     fn build (&self, app: &mut App) {
         app
+        .add_event::<MovementComplete>()
         .add_systems(OnEnter(AppState::Playing), distribute_on_grid.in_set(ApplyGridMovement))
         .add_systems(Update, (
             snap_to_grid,
@@ -25,13 +26,27 @@ pub struct GridLocation(pub IVec2);
 #[derive(Component)]
 pub struct SnapToGrid;
 
-pub fn center_of(grid_location: &GridLocation) -> Vec2 {
+#[derive(Event)]
+pub struct MovementComplete {
+    pub entity: Entity,
+}
+
+fn center_of(grid_location: &GridLocation) -> Vec2 {
     Vec2::new((grid_location.x * GRID_SPACING) as f32, (grid_location.y * GRID_SPACING) as f32)
 }
 
-fn snap_to_grid(mut query: Query<(&mut Transform, Option<&mut AnimateTranslation>, &GridLocation), (With<SnapToGrid>, Changed<GridLocation>)>) {
+fn snap_to_grid(
+    mut query: Query<(&mut Transform, Option<&mut AnimateTranslation>, Ref<GridLocation>),
+    (With<SnapToGrid>, Changed<GridLocation>)>
+) {
     for (mut transform, animate_transform, grid_location) in query.iter_mut() {
-        let destination = Vec2::new((grid_location.x * GRID_SPACING) as f32, (grid_location.y * GRID_SPACING) as f32);
+        let destination = center_of(&grid_location);
+        // Insta-snap newly added components.
+        if grid_location.is_added() {
+            transform.translation = destination.extend(0.);
+            continue;
+        }
+
         match animate_transform {
             Some(mut animate_transform) => {
                 animate_transform.start = transform.translation.truncate();
@@ -89,10 +104,19 @@ pub struct AnimateTranslation {
     pub ease: CubicSegment<Vec2>
 }
 
-fn animate_translation(time: Res<Time>, mut query: Query<(&mut Transform, &mut AnimateTranslation)>) {
-    for (mut transform, mut animate_translation) in query.iter_mut() {
+fn animate_translation(
+    time: Res<Time>,
+    mut event_writer: EventWriter<MovementComplete>,
+    mut query: Query<(Entity, &mut Transform, &mut AnimateTranslation)>
+) {
+    for (entity, mut transform, mut animate_translation) in query.iter_mut() {
+        if animate_translation.timer.finished() {
+            continue;
+        }
+
         if animate_translation.timer.tick(time.delta()).just_finished() {
             transform.translation = animate_translation.end.extend(0.);
+            event_writer.send(MovementComplete{entity});
         } else {
             let progress = animate_translation.timer.percent();
             let lerp = animate_translation.ease.ease(progress);
